% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CubeSum.R
\name{CubeSum}
\alias{CubeSum}
\title{Compute sum of cubes 1^3 + ... + n^3}
\usage{
CubeSum(n)
}
\arguments{
\item{n}{Integer, upper limit}
}
\value{
Sum of cubes
}
\description{
Compute sum of cubes 1^3 + ... + n^3
}
